Initial bezier curve
This is the initial revision of the bezier curve function created within PICO8. The bezier curve and its bounding block can be created within the software, scaling exists with gridlines. User interaction needs to be added such as the ability to increase the scale during the programs runtime, and initial ability to specify the control points for the bezier curve.